% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERdfEXTRACTDS.R
\name{DANGERdfEXTRACTDS}
\alias{DANGERdfEXTRACTDS}
\title{DANGER function returning data.frame (df) from processing servers to
clientside - *** DEPRECATED DANGERdmtEXTRACTDS ***}
\usage{
DANGERdfEXTRACTDS(dataframeName = NULL)
}
\arguments{
\item{dataframeName}{character string that specifies the name of the df 
to be copied from the serverside to the clientside}
}
\value{
the serverside df specified by name dataframeName as a
data.frame/matrix/tibble on the clientside.
}
\description{
DANGERdfEXTRACTDS copies a df from data servers to clientside.
}
\details{
*** DEPRECATED BY DANGERdmtEXTRACTDS ***
}
\author{
DataSHIELD Development Team
}
