% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERdmtEXTRACTDS.R
\name{DANGERdmtEXTRACTDS}
\alias{DANGERdmtEXTRACTDS}
\title{DANGER function returning data.frame, Matrix or tibble (DMT) from
processing servers to clientside}
\usage{
DANGERdmtEXTRACTDS(dfName)
}
\arguments{
\item{dfName}{is a character string that specifies the name of the DMT 
to be copied from the serverside to the clientside. Fully speciifed
by <dfName> argument in ds.DANGERdmtEXTRACT}
}
\value{
the serverside dmt specified by name dfName as a
data.frame/matrix/tibble on the clientside.
}
\description{
DANGERdmtEXTRACTDS copies a DMT from data servers to clientside.
}
\details{
Serverside aggregate function DANGERdmtEXTRACTDS
called by ds.DANGERdmtEXTRACT. DANGER FUNCTION: must not be
installed on data processing servers holding
real data. ONLY TO BE USED FOR DEVELOPMENT PURPOSES.
For more information about how it works see help for ds.DANGERdmtEXTRACT
}
\author{
Paul Burton for DataSHIELD Development Team - 4th June, 2021
}
