% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dangerDS.R
\name{dangerDS}
\alias{dangerDS}
\title{Returns any object to the client side}
\usage{
dangerDS(x)
}
\arguments{
\item{x}{the name of the server side object that will be returned}
}
\value{
a numeric, the statistical mean
}
\description{
DANGER - special function to allow analysts to look at server side objects
on the client side
}
\details{


this function is not meant to be used anywhere except in a development environment
It is designed to give the analyst access to the data (i.e. a route through the DataSHIELD
defences)
}
\author{
Bishop, T.
}

