% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERlistcodeDS.R
\name{DANGERlistcodeDS}
\alias{DANGERlistcodeDS}
\title{lists the code for a function}
\usage{
DANGERlistcodeDS(package.name = NULL, function.name = NULL)
}
\arguments{
\item{package.name}{a vector of character containing a package name}

\item{function.name}{a vector of character containing a function name}
}
\value{
a vector of character containing the code for the function, or a stop error message
}
\description{
list the code for a named fucntion in a named package
}
\details{
Serverside aggregate function \code{DANGERsearchDS} called by clientside function
}
\author{
Stuart Wheater, DataSHIELD Team.
}
