% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERpassPARSERDS.R
\name{DANGERpassPARSERDS}
\alias{DANGERpassPARSERDS}
\title{determines how a given R object or character string is managed by the DataSHIELD parser}
\usage{
DANGERpassPARSERDS(transmit.object)
}
\arguments{
\item{transmit.object}{fully specified by equivalent clientside argument.}
}
\value{
information about whether the input object/string gets blocked or passed
by the parser. If it passes the parser, the object itself and a record of its class
are passed back to the clientside.
}
\description{
presents an R object or character string to the DataSHIELD R parser
and determines whether it is passed by the parser and, if so, what form it
takes after it has traversed the parser.
}
\details{
Serverside aggregate function {DANGERpassPARSERDS} called by clientside function
{ds.DANGERpassPARSER} which presents an R object or character string to the DataSHIELD R parser
and determines whether it is passed or blocked by the parser. If it passes,
it then explores what form it takes after it has traversed the parser. For further details
see help for {ds.DANGERpassPARSER}.
}
\author{
Paul Burton (2020)
}
