% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERplotDS.R
\name{DANGERplotDS}
\alias{DANGERplotDS}
\title{DANGERplotDS}
\usage{
DANGERplotDS(x.name = NULL, y.name = NULL)
}
\arguments{
\item{x.name}{character string naming first object to be
plotted on clientside
from serverside to clientside}

\item{y.name}{character string naming second object to be
plotted on clientside}
}
\value{
DANGER function
}
\description{
DANGERplotDS
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
