% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERprintDS.R
\name{DANGERprintDS}
\alias{DANGERprintDS}
\title{DANGERprintDS}
\usage{
DANGERprintDS(xName = NULL, yName = NULL, zName = NULL)
}
\arguments{
\item{xName}{character string naming first object to be
printed on clientside
from serverside to clientside}

\item{yName}{character string naming second object to be
printed on clientside}

\item{zName}{character string naming third object to be 
printed on clientside}
}
\value{
DANGER function
}
\description{
DANGERprintDS
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
