% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERsearchDS.R
\name{DANGERsearchDS}
\alias{DANGERsearchDS}
\title{lists all environments on serverside}
\usage{
DANGERsearchDS()
}
\value{
a vector of character strings to the clientside giving the names of
each of the R environments on the serverside in order of their numeric "pos"
}
\description{
creates a list of the names of all of the serverside
environments
}
\details{
Serverside aggregate function {DANGERsearchDS} called by clientside function
{ds.DANGERsearch}. No arguments need to be specified
}
\author{
Paul Burton 2020
}
