% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERseedEXTRACTDS.R
\name{DANGERseedEXTRACTDS}
\alias{DANGERseedEXTRACTDS}
\title{DANGERseedEXTRACT}
\usage{
DANGERseedEXTRACTDS()
}
\value{
DANGER function
}
\description{
DANGERseedEXTRACT
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
