% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANGERvarsEXTRACTDS.R
\name{DANGERvarsEXTRACTDS}
\alias{DANGERvarsEXTRACTDS}
\title{DANGERvarsEXTRACT}
\usage{
DANGERvarsEXTRACTDS(
  xName = NULL,
  yName = NULL,
  zName = NULL,
  uName = NULL,
  vName = NULL,
  wName = NULL
)
}
\arguments{
\item{xName}{character string naming first variable to be copied
from serverside to clientside}

\item{yName}{character string naming second variable to be copied
from serverside to clientside}

\item{zName}{character string naming third variable to be copied
from serverside to clientside}

\item{uName}{character string naming fourth variable to be copied
from serverside to clientside}

\item{vName}{character string naming fifth variable to be copied
from serverside to clientside}

\item{wName}{character string naming sixth variable to be copied
from serverside to clientside}
}
\value{
DANGER function
}
\description{
DANGERvarsEXTRACT
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
